refactor(tui): split the 3.6k-line app.rs into concern modules
Behavior-preserving decomposition of the monolithic App impl. app.rs becomes
app/mod.rs (struct, types, constructor, small helpers) plus one `impl App`
submodule per concern:

- app/input.rs    key dispatch + all handle_*_key + enter_*_mode (was the giant handle_key tree)
- app/update.rs   the lifecycle loop — refresh (457 LoC), tick, run_auto_actions, budgets, coord
- app/actions.rs  session actions (approve, brain accept/reject, toggle, compact, switch)
- app/overlays.rs skills / brain / hive overlay open + refresh
- app/filters.rs  status/focus/search filtering + visible-session selection
- app/demo.rs     refresh_demo

Cross-submodule private methods promoted to pub(super) (visible within the app
subtree only). No public API change; 20 app tests pass unchanged; clippy
-D warnings clean.
