+++ crates/claudectl-tui/src/app/update.rs
@@ -1,0 +1,925 @@
+//! Extracted from app/mod.rs — behavior-preserving split.
+#![allow(clippy::too_many_lines)]
+
+use super::*;
+
+impl App {
+    pub fn refresh(&mut self) {
+        let tick_start = std::time::Instant::now();
+
+        if self.demo_mode {
+            self.refresh_demo();
+            if self.debug {
+                let total_elapsed = tick_start.elapsed();
+                self.debug_timings
+                    .record(0.0, 0.0, 0.0, total_elapsed.as_secs_f64() * 1000.0);
+            }
+            return;
+        }
+
+        // Discover which PIDs have session files
+        let scan_start = std::time::Instant::now();
+        let discovered = discovery::scan_sessions();
+        let scan_elapsed = scan_start.elapsed();
+
+        // Build a map of existing sessions by PID for state preservation
+        let mut existing: HashMap<u32, ClaudeSession> =
+            self.sessions.drain(..).map(|s| (s.pid, s)).collect();
+
+        // Merge: reuse existing session state (jsonl_offset, tokens, cost, cpu_history)
+        // or create new from discovered
+        let mut new_pids: Vec<u32> = Vec::new();
+        let mut sessions: Vec<ClaudeSession> = discovered
+            .into_iter()
+            .map(|new| {
+                if let Some(mut prev) = existing.remove(&new.pid) {
+                    // Preserve accumulated state, update ephemeral fields
+                    prev.elapsed = new.elapsed;
+                    prev.started_at = new.started_at;
+                    // cwd/project_name/session_id don't change
+                    prev
+                } else {
+                    // Brand new session
+                    new_pids.push(new.pid);
+                    new
+                }
+            })
+            .collect();
+
+        // Enrich with ps data (CPU, MEM, TTY, command args) + filter dead PIDs
+        let ps_start = std::time::Instant::now();
+        process::fetch_and_enrich(&mut sessions);
+        let ps_elapsed = ps_start.elapsed();
+
+        // Resolve JSONL paths (only for sessions that don't have one yet)
+        for session in &mut sessions {
+            if session.jsonl_path.is_none() {
+                discovery::resolve_jsonl_paths(std::slice::from_mut(session));
+            }
+        }
+
+        // Scan for subagents
+        discovery::scan_subagents(&mut sessions);
+
+        // Resolve git worktree identity (for conflict detection, runs once per session)
+        discovery::resolve_worktree_ids(&mut sessions);
+
+        // Snapshot previous cost for burn rate BEFORE reading new JSONL data
+        for session in &mut sessions {
+            session.prev_cost_usd = session.cost_usd;
+        }
+
+        // Read JSONL incrementally (only new bytes since last offset)
+        let jsonl_start = std::time::Instant::now();
+        for session in &mut sessions {
+            monitor::update_tokens(session);
+        }
+        let jsonl_elapsed = jsonl_start.elapsed();
+
+        // Compute burn rate from cost delta (skip first tick where prev_cost is 0)
+        for session in &mut sessions {
+            if session.prev_cost_usd > 0.001 {
+                let delta = session.cost_usd - session.prev_cost_usd;
+                if delta > 0.001 {
+                    session.burn_rate_per_hr = delta * 1800.0;
+                } else {
+                    // Decay burn rate toward zero when no new cost
+                    session.burn_rate_per_hr *= 0.5;
+                    if session.burn_rate_per_hr < 0.01 {
+                        session.burn_rate_per_hr = 0.0;
+                    }
+                }
+            }
+            // Fold the instantaneous rate into the smoothed EWMA + sample window
+            // used for budget forecasting (#370). The 2s constant matches the
+            // `* 1800.0` per-hour conversion above (3600 / 1800 = 2s ticks).
+            session.record_burn_sample(2.0);
+        }
+
+        // Budget enforcement
+        if let Some(budget) = self.budget_usd {
+            for session in &sessions {
+                let pct = session.cost_usd / budget * 100.0;
+
+                // Warn at 80%
+                if (80.0..100.0).contains(&pct) && !self.budget_warned.contains(&session.pid) {
+                    self.budget_warned.insert(session.pid);
+                    self.status_msg = format!(
+                        "BUDGET WARNING: {} at {:.0}% (${:.2}/${:.2})",
+                        session.display_name(),
+                        pct,
+                        session.cost_usd,
+                        budget
+                    );
+                    self.notify_user(
+                        &format!("budget-warn:{}", session.pid),
+                        &format!("{} budget at {:.0}%", session.display_name(), pct),
+                    );
+                    self.hooks.fire(HookEvent::BudgetWarning, session);
+                }
+
+                // Kill at 100%
+                if pct >= 100.0 && !self.budget_killed.contains(&session.pid) {
+                    self.budget_killed.insert(session.pid);
+                    if self.kill_on_budget {
+                        let _ = self.runtime.actions.terminate_session(session.pid);
+                        self.status_msg = format!(
+                            "BUDGET EXCEEDED: Killed {} (${:.2}/${:.2})",
+                            session.display_name(),
+                            session.cost_usd,
+                            budget
+                        );
+                    } else {
+                        self.status_msg = format!(
+                            "BUDGET EXCEEDED: {} at ${:.2}/{:.2} — use --kill-on-budget to auto-kill",
+                            session.display_name(),
+                            session.cost_usd,
+                            budget
+                        );
+                    }
+                    self.notify_user(
+                        &format!("budget-kill:{}", session.pid),
+                        &format!("{} exceeded budget!", session.display_name()),
+                    );
+                    self.hooks.fire(HookEvent::BudgetExceeded, session);
+                }
+            }
+        }
+
+        // Context threshold warnings
+        if self.context_warn_threshold > 0 {
+            let threshold = self.context_warn_threshold as f64;
+            for session in &sessions {
+                let pct = session.context_percent();
+                if pct >= threshold && !self.context_warned.contains(&session.pid) {
+                    self.context_warned.insert(session.pid);
+                    self.status_msg = format!(
+                        "CONTEXT HIGH: {} at {:.0}% of context window",
+                        session.display_name(),
+                        pct
+                    );
+                    self.notify_user(
+                        &format!("context:{}", session.pid),
+                        &format!("{} context at {:.0}%", session.display_name(), pct),
+                    );
+                    self.hooks.fire(HookEvent::ContextHigh, session);
+                } else if pct < threshold && self.context_warned.contains(&session.pid) {
+                    // Reset warning if context dropped (e.g., after /compact)
+                    self.context_warned.remove(&session.pid);
+                }
+            }
+        }
+
+        // Record activity for sparkline and cache decay score
+        for session in &mut sessions {
+            session.record_activity();
+            session.decay_score =
+                claudectl_core::health::compute_decay_score(session, &self.health_thresholds);
+        }
+
+        // Track when sessions first appear as Finished, remove after 30s
+        let now = std::time::Instant::now();
+        for session in &sessions {
+            if session.status == SessionStatus::Finished
+                && !self.finished_at.contains_key(&session.pid)
+            {
+                self.finished_at.insert(session.pid, now);
+                // Record to history on first Finished detection
+                claudectl_core::history::record_session(session);
+            }
+        }
+        sessions.retain(|s| {
+            if s.status == SessionStatus::Finished {
+                if let Some(&t) = self.finished_at.get(&s.pid) {
+                    return now.duration_since(t).as_secs() < 30;
+                }
+            }
+            true
+        });
+        // Clean up old finished_at entries + their session files
+        let expired: Vec<u32> = self
+            .finished_at
+            .iter()
+            .filter(|(_, t)| now.duration_since(**t).as_secs() >= 60)
+            .map(|(pid, _)| *pid)
+            .collect();
+        for pid in &expired {
+            let session_file = dirs_home()
+                .join(".claude/sessions")
+                .join(format!("{pid}.json"));
+            let _ = std::fs::remove_file(session_file);
+        }
+        self.finished_at
+            .retain(|_, t| now.duration_since(*t).as_secs() < 60);
+
+        // Sort
+        self.apply_sort(&mut sessions);
+
+        // Notifications and webhooks: check for status transitions
+        for session in &sessions {
+            let prev = self.prev_statuses.get(&session.pid).copied();
+            let changed = prev.is_some() && prev != Some(session.status);
+
+            if !changed {
+                continue;
+            }
+
+            claudectl_core::logger::log(
+                "DEBUG",
+                &format!(
+                    "session {}: status {} -> {}",
+                    session.display_name(),
+                    prev.unwrap(),
+                    session.status
+                ),
+            );
+
+            // Desktop notification on NeedsInput (gated + cooldown via notify_user)
+            if session.status == SessionStatus::NeedsInput {
+                self.notify_user(
+                    &format!("needs-input:{}", session.pid),
+                    &format!("{} needs input", session.project_name),
+                );
+            }
+
+            // Webhook on status change
+            if let Some(ref url) = self.webhook_url {
+                let new_status = session.status.to_string();
+                let should_fire = match &self.webhook_filter {
+                    Some(filter) => filter.iter().any(|f| f.eq_ignore_ascii_case(&new_status)),
+                    None => true,
+                };
+                if should_fire {
+                    claudectl_core::logger::log(
+                        "DEBUG",
+                        &format!(
+                            "webhook fired for {} -> {}",
+                            session.display_name(),
+                            new_status
+                        ),
+                    );
+                    fire_webhook(
+                        url,
+                        session,
+                        prev.map(|p| p.to_string()).unwrap_or_default(),
+                    );
+                }
+            }
+
+            // Event hooks
+            self.hooks.fire_with_status(
+                HookEvent::StatusChange,
+                session,
+                &prev.unwrap().to_string(),
+                &session.status.to_string(),
+            );
+
+            match session.status {
+                SessionStatus::NeedsInput => {
+                    self.hooks.fire(HookEvent::NeedsInput, session);
+                }
+                SessionStatus::Finished => {
+                    self.hooks.fire(HookEvent::Finished, session);
+                }
+                SessionStatus::Idle => {
+                    self.hooks.fire(HookEvent::Idle, session);
+                }
+                _ => {}
+            }
+        }
+
+        // Fire hooks for newly discovered sessions
+        for session in sessions.iter().filter(|s| new_pids.contains(&s.pid)) {
+            self.hooks.fire(HookEvent::SessionStart, session);
+        }
+
+        // Track NeedsInput wait times
+        let now_instant = std::time::Instant::now();
+        for session in &sessions {
+            if session.status == SessionStatus::NeedsInput {
+                // Record when it first entered NeedsInput
+                self.needs_input_since
+                    .entry(session.pid)
+                    .or_insert(now_instant);
+            } else {
+                // Clear if no longer NeedsInput
+                self.needs_input_since.remove(&session.pid);
+            }
+        }
+        // Clean up entries for sessions that no longer exist
+        let active_pids: HashSet<u32> = sessions.iter().map(|s| s.pid).collect();
+        self.needs_input_since
+            .retain(|pid, _| active_pids.contains(pid));
+
+        // Conflict detection: find sessions sharing the same git worktree
+        // Uses worktree_id (git show-toplevel) so different worktrees don't false-positive
+        self.conflict_pids.clear();
+        let mut wt_sessions: HashMap<&str, Vec<u32>> = HashMap::new();
+        for session in &sessions {
+            if session.status != SessionStatus::Finished {
+                let key = session.worktree_id.as_deref().unwrap_or(&session.cwd);
+                wt_sessions.entry(key).or_default().push(session.pid);
+            }
+        }
+        for (wt, pids) in &wt_sessions {
+            if pids.len() >= 2 {
+                for &pid in pids {
+                    self.conflict_pids.insert(pid);
+                }
+                // Fire hook once per worktree conflict (not on every tick)
+                if !self.conflict_alerted.contains(*wt) {
+                    self.conflict_alerted.insert(wt.to_string());
+                    let project = sessions
+                        .iter()
+                        .find(|s| s.pid == pids[0])
+                        .map(|s| s.display_name())
+                        .unwrap_or("unknown");
+                    self.status_msg =
+                        format!("CONFLICT: {} sessions sharing {}", pids.len(), project);
+                    self.notify_user(
+                        &format!("conflict:{wt}"),
+                        &format!("{} sessions in {}", pids.len(), project),
+                    );
+                    if let Some(session) = sessions.iter().find(|s| s.pid == pids[0]) {
+                        self.hooks.fire(HookEvent::ConflictDetected, session);
+                    }
+                }
+            }
+        }
+        // Clear alerts for worktrees that no longer have conflicts
+        self.conflict_alerted.retain(|wt| {
+            wt_sessions
+                .get(wt.as_str())
+                .map(|pids| pids.len() >= 2)
+                .unwrap_or(false)
+        });
+
+        // File-level conflict detection: find files edited by multiple sessions
+        self.file_conflict_pids.clear();
+        self.file_conflicts.clear();
+        // Reset has_file_conflict on all sessions
+        for session in &mut sessions {
+            session.has_file_conflict = false;
+        }
+
+        if self.file_conflicts_enabled {
+            // Build file → PIDs map from files_modified across active sessions
+            let mut file_pids: HashMap<String, Vec<u32>> = HashMap::new();
+            for session in &sessions {
+                if session.status == SessionStatus::Finished {
+                    continue;
+                }
+                for file in session.files_modified.keys() {
+                    file_pids.entry(file.clone()).or_default().push(session.pid);
+                }
+                // Also consider pending file edits (predictive conflict)
+                if let Some(ref pending) = session.pending_file_path {
+                    file_pids
+                        .entry(pending.clone())
+                        .or_default()
+                        .push(session.pid);
+                }
+            }
+
+            // Deduplicate PIDs per file (a session may appear twice if it both modified and is pending)
+            for pids in file_pids.values_mut() {
+                pids.sort_unstable();
+                pids.dedup();
+            }
+
+            // Record conflicts where 2+ sessions touch the same file
+            for (file, pids) in &file_pids {
+                if pids.len() >= 2 {
+                    for &pid in pids {
+                        self.file_conflict_pids.insert(pid);
+                    }
+                    self.file_conflicts.insert(file.clone(), pids.clone());
+
+                    // Mark sessions with pending file conflicts
+                    for session in &mut sessions {
+                        if let Some(ref pending) = session.pending_file_path {
+                            if pending == file && pids.contains(&session.pid) {
+                                session.has_file_conflict = true;
+                            }
+                        }
+                    }
+
+                    // Fire alert once per conflicting file
+                    if !self.file_conflict_alerted.contains(file) {
+                        self.file_conflict_alerted.insert(file.clone());
+                        let names: Vec<&str> = pids
+                            .iter()
+                            .filter_map(|pid| {
+                                sessions
+                                    .iter()
+                                    .find(|s| s.pid == *pid)
+                                    .map(|s| s.display_name())
+                            })
+                            .collect();
+                        let short = file.rsplit('/').next().unwrap_or(file);
+                        self.status_msg =
+                            format!("FILE CONFLICT: {} edited by {}", short, names.join(", "));
+                        self.notify_user(
+                            &format!("file-conflict:{file}"),
+                            &format!("File conflict: {short}"),
+                        );
+                        if let Some(session) = sessions.iter().find(|s| s.pid == pids[0]) {
+                            self.hooks.fire(HookEvent::ConflictDetected, session);
+                        }
+                    }
+                }
+            }
+
+            // Clear alerts for files no longer in conflict
+            self.file_conflict_alerted
+                .retain(|f| self.file_conflicts.contains_key(f));
+        }
+
+        // Update prev_statuses
+        self.prev_statuses = sessions.iter().map(|s| (s.pid, s.status)).collect();
+
+        self.sessions = sessions;
+
+        // Append remote sessions from relay peers (if relay feature active)
+        #[cfg(feature = "relay")]
+        {
+            for remote in &self.remote_sessions {
+                self.sessions.push(remote.clone());
+            }
+        }
+
+        self.normalize_selection();
+
+        // Record debug timings
+        if self.debug {
+            let total_elapsed = tick_start.elapsed();
+            self.debug_timings.record(
+                scan_elapsed.as_secs_f64() * 1000.0,
+                ps_elapsed.as_secs_f64() * 1000.0,
+                jsonl_elapsed.as_secs_f64() * 1000.0,
+                total_elapsed.as_secs_f64() * 1000.0,
+            );
+        }
+    }
+
+    pub fn tick(&mut self) {
+        self.status_msg.clear();
+
+        let now_ms = std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .unwrap_or_default()
+            .as_millis() as u64;
+
+        for session in &mut self.sessions {
+            let elapsed_ms = now_ms.saturating_sub(session.started_at);
+            session.elapsed = std::time::Duration::from_millis(elapsed_ms);
+        }
+
+        self.refresh();
+        self.run_auto_actions();
+
+        // Check idle mode transition
+        self.check_idle_mode();
+
+        // Refresh weekly summary every ~30s (15 ticks at 2s interval)
+        self.weekly_summary_tick += 1;
+        if self.weekly_summary_tick >= 15 {
+            self.weekly_summary_tick = 0;
+            self.weekly_summary = claudectl_core::history::weekly_summary();
+            self.check_aggregate_budgets();
+        }
+
+        // Refresh coordination state every ~6s (3 ticks at 2s interval)
+        #[cfg(feature = "coord")]
+        {
+            self.coord_tick += 1;
+            if self.coord_tick >= 3 {
+                self.coord_tick = 0;
+                self.coord_refresh();
+            }
+        }
+    }
+
+    pub(super) fn run_auto_actions(&mut self) {
+        // In demo mode, events are scripted in refresh_demo() — skip real execution
+        if self.demo_mode {
+            return;
+        }
+
+        // Legacy per-PID auto-approve (toggled with 'a' key)
+        let legacy_pids: Vec<u32> = self
+            .sessions
+            .iter()
+            .filter(|s| s.status == SessionStatus::NeedsInput && self.auto_approve.contains(&s.pid))
+            .map(|s| s.pid)
+            .collect();
+
+        for pid in legacy_pids {
+            if let Some(session) = self.sessions.iter().find(|s| s.pid == pid) {
+                let _ = self
+                    .runtime
+                    .actions
+                    .log_observation(observation_from(session, "user_approve"));
+                match terminals::approve_session(session) {
+                    Ok(()) => self.status_msg = format!("Auto-approved {}", session.display_name()),
+                    Err(e) => self.status_msg = format!("Auto-approve error: {e}"),
+                }
+            }
+        }
+
+        // Built-in file conflict auto-deny: deny writes to files being edited by another session
+        if self.auto_deny_file_conflicts {
+            let conflict_candidates: Vec<(u32, String, String)> = self
+                .sessions
+                .iter()
+                .filter(|s| {
+                    s.status == SessionStatus::NeedsInput
+                        && s.has_file_conflict
+                        && s.pending_file_path.is_some()
+                })
+                .filter_map(|s| {
+                    let file = s.pending_file_path.as_ref()?;
+                    let other_pids = self.file_conflicts.get(file)?;
+                    let other_name = other_pids
+                        .iter()
+                        .filter(|&&p| p != s.pid)
+                        .find_map(|pid| {
+                            self.sessions
+                                .iter()
+                                .find(|o| o.pid == *pid)
+                                .map(|o| format!("{} (PID {})", o.display_name(), o.pid))
+                        })
+                        .unwrap_or_else(|| "another session".into());
+                    Some((s.pid, file.clone(), other_name))
+                })
+                .collect();
+
+            for (pid, file, other) in conflict_candidates {
+                // Debounce
+                if let Some(last) = self.auto_actions_fired.get(&pid) {
+                    if last.elapsed().as_secs() < 5 {
+                        continue;
+                    }
+                }
+                if let Some(session) = self.sessions.iter().find(|s| s.pid == pid) {
+                    // Log passive observation: conflict auto-deny
+                    let _ = self
+                        .runtime
+                        .actions
+                        .log_observation(observation_from(session, "conflict_deny"));
+                    let short = file.rsplit('/').next().unwrap_or(&file);
+                    let msg = format!("File {short} is being edited by {other}");
+                    match self.runtime.actions.inject_text(&session.session_id, &msg) {
+                        Ok(()) => {
+                            let status = format!(
+                                "File conflict: denied {} edit to {short}",
+                                session.display_name()
+                            );
+                            claudectl_core::logger::log("CONFLICT", &status);
+                            self.status_msg = status;
+                        }
+                        Err(e) => {
+                            self.status_msg = format!("File conflict deny error: {e}");
+                        }
+                    }
+                    self.auto_actions_fired
+                        .insert(pid, std::time::Instant::now());
+                }
+            }
+        }
+
+        // Rule-based auto-actions
+        if !self.rules.is_empty() {
+            let candidates: Vec<u32> = self
+                .sessions
+                .iter()
+                .filter(|s| {
+                    matches!(
+                        s.status,
+                        SessionStatus::NeedsInput | SessionStatus::WaitingInput
+                    )
+                })
+                .filter(|s| !self.auto_approve.contains(&s.pid)) // Legacy takes priority
+                .map(|s| s.pid)
+                .collect();
+
+            for pid in candidates {
+                // Debounce: don't re-fire within 3 seconds for same PID
+                if let Some(last) = self.auto_actions_fired.get(&pid) {
+                    if last.elapsed().as_secs() < 3 {
+                        continue;
+                    }
+                }
+
+                let session = match self.sessions.iter().find(|s| s.pid == pid) {
+                    Some(s) => s,
+                    None => continue,
+                };
+
+                let result = claudectl_core::rules::evaluate(&self.rules, session);
+                let Some(rule_match) = result else {
+                    continue;
+                };
+
+                // Log passive observation: static rule fired
+                let obs_action = format!("rule_{}", rule_match.action.label());
+                let _ = self
+                    .runtime
+                    .actions
+                    .log_observation(observation_from(session, &obs_action));
+
+                let msg = claudectl_core::rules::execute(&rule_match, session);
+                match msg {
+                    Ok(status) => {
+                        claudectl_core::logger::log("AUTO", &status);
+                        self.last_rule_action = Some(status.clone());
+                        self.status_msg = status;
+                    }
+                    Err(e) => {
+                        self.status_msg = format!("Rule error: {e}");
+                    }
+                }
+
+                self.auto_actions_fired
+                    .insert(pid, std::time::Instant::now());
+            }
+        } // end if !self.rules.is_empty()
+
+        // Brain inference (opt-in, runs after rules)
+        if let Some(ref mut driver) = self.brain_driver {
+            // Collect deny-only rules for override checking
+            let deny_rules: Vec<_> = self
+                .rules
+                .iter()
+                .filter(|r| r.action == claudectl_core::rules::RuleAction::Deny)
+                .cloned()
+                .collect();
+
+            let snapshots: Vec<_> = self.sessions.iter().map(snapshot_from).collect();
+            let actions = driver.tick(&snapshots, &deny_rules);
+            for (_pid, msg) in actions {
+                claudectl_core::logger::log("BRAIN", &msg);
+                self.status_msg = msg;
+            }
+
+            driver.cleanup(&snapshots);
+
+            // Deliver pending mailbox messages to sessions waiting for input.
+            // The orchestrator resolves SessionSnapshot back to live sessions
+            // internally; we project once here.
+            let snapshots: Vec<_> = self.sessions.iter().map(snapshot_from).collect();
+            let deliveries = self.runtime.orchestrator.deliver_mailbox(&snapshots);
+            for (_pid, msg) in deliveries {
+                claudectl_core::logger::log("MAILBOX", &msg);
+                self.status_msg = msg;
+            }
+        }
+
+        // Deliver pending typed interrupts from the coordination bus. The
+        // orchestrator handles the SQLite connection internally.
+        #[cfg(feature = "coord")]
+        {
+            let snapshots: Vec<_> = self.sessions.iter().map(snapshot_from).collect();
+            let deliveries = self.runtime.orchestrator.deliver_interrupts(&snapshots);
+            for (_intr_id, msg) in deliveries {
+                claudectl_core::logger::log("INTERRUPT", &msg);
+                self.status_msg = msg;
+            }
+        }
+    }
+
+    pub fn handle_auto_approve(&mut self) {
+        let Some(session) = self.selected_session() else {
+            return;
+        };
+        if session.is_remote() {
+            self.status_msg = "Remote session \u{2014} action not available".into();
+            return;
+        }
+        let pid = session.pid;
+        let name = session.display_name().to_string();
+
+        if self.pending_auto_approve == Some(pid) {
+            if self.auto_approve.contains(&pid) {
+                self.auto_approve.remove(&pid);
+                self.status_msg = format!("Auto-approve OFF for {name}");
+            } else {
+                self.auto_approve.insert(pid);
+                self.status_msg = format!("Auto-approve ON for {name}");
+            }
+            self.pending_auto_approve = None;
+        } else {
+            self.pending_auto_approve = Some(pid);
+            let action = if self.auto_approve.contains(&pid) {
+                "disable"
+            } else {
+                "enable"
+            };
+            self.status_msg = format!("Press a again to {action} auto-approve for {name}");
+        }
+    }
+
+    pub fn handle_kill(&mut self) {
+        let Some(session) = self.selected_session() else {
+            return;
+        };
+        if session.is_remote() {
+            self.status_msg = "Remote session \u{2014} action not available".into();
+            return;
+        }
+        let pid = session.pid;
+        let name = session.display_name().to_string();
+
+        if self.pending_kill == Some(pid) {
+            match self.runtime.actions.terminate_session(pid) {
+                Ok(()) => {
+                    self.status_msg = format!("Killed {name} (PID {pid})");
+                    self.auto_approve.remove(&pid);
+                    // Don't delete session file yet — let the Finished tombstone show for 30s.
+                    // The file will be cleaned up when the tombstone expires.
+                    self.refresh();
+                }
+                Err(e) => self.status_msg = format!("Kill failed: {e}"),
+            }
+            self.pending_kill = None;
+        } else {
+            self.pending_kill = Some(pid);
+            self.status_msg = format!("Kill {name} (PID {pid})? Press d again to confirm");
+        }
+    }
+
+    pub fn cancel_pending_kill(&mut self) {
+        if self.pending_kill.is_some() {
+            self.pending_kill = None;
+            self.status_msg = "Kill cancelled".into();
+        }
+    }
+
+    /// Refresh cached coordination state from the runtime.
+    ///
+    /// The `expire_stale_*` calls remain direct because they're side-effects
+    /// on the SQLite store (bookkeeping), not part of the read-only
+    /// `CoordView` surface. The actual list queries go through the runtime
+    /// trait so the binary-coord coupling stays one layer thick.
+    #[cfg(feature = "coord")]
+    pub fn coord_refresh(&mut self) {
+        // Bookkeeping (expire stale leases + interrupts). Best-effort; the
+        // orchestrator logs failures internally and never propagates them.
+        self.runtime.orchestrator.expire_stale();
+
+        self.coord_leases = self.runtime.coord.active_leases();
+        self.coord_handoffs = self.runtime.coord.pending_handoffs();
+        self.coord_pending_interrupts = self.runtime.coord.pending_interrupts();
+        self.coord_tasks = self.runtime.coord.tasks();
+        if self.supervisor_selected >= self.coord_tasks.len() {
+            self.supervisor_selected = self.coord_tasks.len().saturating_sub(1);
+        }
+        self.coord_task_sessions = self
+            .coord_tasks
+            .iter()
+            .filter_map(|t| t.last_session_id.clone())
+            .collect();
+
+        self.coord_lease_sessions = self
+            .coord_leases
+            .iter()
+            .map(|l| l.owner_session_id.clone())
+            .collect();
+        self.coord_handoff_sessions = self
+            .coord_handoffs
+            .iter()
+            .flat_map(|h| {
+                let mut ids = vec![h.from_session_id.clone()];
+                if let Some(ref to) = h.to_session_id {
+                    ids.push(to.clone());
+                }
+                ids
+            })
+            .collect();
+        self.coord_interrupt_targets = self
+            .coord_pending_interrupts
+            .iter()
+            .map(|i| i.target_session_id.clone())
+            .collect();
+    }
+
+    pub(super) fn check_aggregate_budgets(&mut self) {
+        // Copy the scalar totals up front so the immutable borrow of
+        // `weekly_summary` doesn't conflict with `notify_user(&mut self)` below.
+        let today_cost_usd = self.weekly_summary.today_cost_usd;
+        let week_cost_usd = self.weekly_summary.cost_usd;
+
+        // Also include cost from currently live sessions (not yet in history)
+        let live_cost: f64 = self.sessions.iter().map(|s| s.cost_usd).sum();
+
+        // Daily limit check
+        if let Some(daily_limit) = self.daily_limit {
+            let today_total = today_cost_usd + live_cost;
+            let pct = today_total / daily_limit * 100.0;
+
+            if pct >= 80.0 && !self.daily_alert_fired {
+                self.daily_alert_fired = true;
+                self.status_msg = format!(
+                    "DAILY BUDGET: ${:.2}/${:.2} ({:.0}%)",
+                    today_total, daily_limit, pct
+                );
+                self.notify_user("daily-budget", &format!("Daily budget at {:.0}%", pct));
+
+                // Fire hooks with a synthetic session containing aggregate data
+                let mut dummy = create_aggregate_session(today_total, daily_limit, "daily");
+                self.hooks.fire(HookEvent::BudgetWarning, &dummy);
+
+                if pct >= 100.0 {
+                    dummy.cost_usd = today_total;
+                    self.hooks.fire(HookEvent::BudgetExceeded, &dummy);
+                }
+            }
+        }
+
+        // Weekly limit check
+        if let Some(weekly_limit) = self.weekly_limit {
+            let week_total = week_cost_usd + live_cost;
+            let pct = week_total / weekly_limit * 100.0;
+
+            if pct >= 80.0 && !self.weekly_alert_fired {
+                self.weekly_alert_fired = true;
+                self.status_msg = format!(
+                    "WEEKLY BUDGET: ${:.2}/${:.2} ({:.0}%)",
+                    week_total, weekly_limit, pct
+                );
+                self.notify_user("weekly-budget", &format!("Weekly budget at {:.0}%", pct));
+
+                let mut dummy = create_aggregate_session(week_total, weekly_limit, "weekly");
+                self.hooks.fire(HookEvent::BudgetWarning, &dummy);
+
+                if pct >= 100.0 {
+                    dummy.cost_usd = week_total;
+                    self.hooks.fire(HookEvent::BudgetExceeded, &dummy);
+                }
+            }
+        }
+    }
+
+    /// Compute budget exhaustion ETA based on current burn rate.
+    /// Returns (spent, limit, eta_string, urgency) where urgency is 0=safe, 1=warn, 2=critical.
+    pub fn budget_eta(&self) -> Option<(f64, f64, String, u8)> {
+        let live_cost: f64 = self.sessions.iter().map(|s| s.cost_usd).sum();
+        let total_burn: f64 = self.sessions.iter().map(|s| s.burn_rate_per_hr).sum();
+
+        // Prefer daily limit, fall back to per-session budget
+        let (spent, limit) = if let Some(daily) = self.daily_limit {
+            (self.weekly_summary.today_cost_usd + live_cost, daily)
+        } else if let Some(budget) = self.budget_usd {
+            // For per-session budget, show the session closest to limit
+            if let Some(session) = self.sessions.iter().max_by(|a, b| {
+                (a.cost_usd / budget)
+                    .partial_cmp(&(b.cost_usd / budget))
+                    .unwrap_or(std::cmp::Ordering::Equal)
+            }) {
+                (session.cost_usd, budget)
+            } else {
+                return None;
+            }
+        } else {
+            return None;
+        };
+
+        let remaining = limit - spent;
+        if remaining <= 0.0 {
+            return Some((spent, limit, "exceeded".into(), 2));
+        }
+        if total_burn < 0.01 {
+            return Some((spent, limit, "safe".into(), 0));
+        }
+
+        let hours_left = remaining / total_burn;
+        let mins_left = (hours_left * 60.0) as u64;
+        let eta_str = if mins_left >= 120 {
+            format!("{}h {}m", mins_left / 60, mins_left % 60)
+        } else {
+            format!("{}m", mins_left)
+        };
+
+        let urgency = if mins_left <= 30 {
+            2
+        } else if mins_left <= 120 {
+            1
+        } else {
+            0
+        };
+        Some((spent, limit, eta_str, urgency))
+    }
+
+    pub(super) fn check_idle_mode(&mut self) {
+        if !self.idle_config.enabled {
+            return;
+        }
+        let idle_threshold = std::time::Duration::from_secs(self.idle_config.after_idle_mins * 60);
+        let was_idle = self.idle_mode_active;
+        self.idle_mode_active = self.last_user_interaction.elapsed() > idle_threshold;
+
+        if self.idle_mode_active && !was_idle {
+            claudectl_core::logger::log("IDLE", "Entering idle mode");
+        }
+    }
+}
